requirements.txt
robotframework>=6.0
robotframework-requests>=0.9.5
robotframework-jsonlibrary>=0.5
flask>=3.0
